chore: Remove deprecated agent repl
