+++ Makefile
@@ -189,9 +189,6 @@ dev.db.console:
 dev.pr.clean.checkout:
 	bash ./scripts/clean-pr-checkout $(PR)
 
-dev.agent.console:
-	$(COMPOSE) exec agent uv run python -m repl.main --interactive --canvas-id "$(SUPERPLANE_CANVAS_ID)" --token "$(SUPERPLANE_API_TOKEN)" --org-id "$(SUPERPLANE_ORG_ID)"
-
 check.db.structure:
 	bash ./scripts/verify_db_structure_clean.sh
