+++ agent/pyproject.toml
@@ -34,7 +34,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/ai", "src/private", "src/repl", "src/superplaneapi", "evals"]
+packages = ["src/ai", "src/private", "src/superplaneapi", "evals"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
