--- agent/src/ai/main.py
@@ -1,7 +1,0 @@
-from repl.main import RELOAD_EXIT_CODE, _parse_stream_event, _stream_repl_answer, main
-
-__all__ = ["RELOAD_EXIT_CODE", "_parse_stream_event", "_stream_repl_answer", "main"]
-
-
-if __name__ == "__main__":
-    main()
