--- agent/src/repl/__init__.py
@@ -1,1 +1,0 @@
-"""REPL CLI package."""
